+++ packages/sub-shared/CHANGELOG.md
@@ -1,5 +1,7 @@
 # @marckrenn/pi-sub-shared
 
+## 1.1.0
+
 ## 1.0.6
 
 ## 1.0.5